Reworked unit tests for better test coverage.
Now uses Test::Fake::HTTPD to verify that the data
input into WebService::Pushover is what we want to
get sent to the remote server. As a side-effect, this
also gives us testing of the API contract that
WebService::Pushover provides to end-users.

This required some changes to Pushover.json, since
the documentation for Net::HTTP::Spore wasn't good enough
to determine how to change per-method base_url values for
testing.