Use core Perl::Critic theme explicitly
This avoids issues when users or testers have other Perl::Critic
policies installed, which might add to the core theme without one
wanting this behaviour.  A good example is the OTRS policy, which adds
several OTRS-specific standards to the default Perl::Critic severity
level and is causing the Perl::Critic warnings as seen in issue #14.
This change implements @olof's recommendation as mentioned in the GitHub
issue.