Define Perl::Critic configuration in an .rc file
Defining the Perl::Critic theme and severity level directly in the
critic.t test suite is the simplest and quickest way of avoiding the
problem of unwanted Perl::Critic policies being used, however it might
be better to define the Perl::Critic configuration in a perlcriticrc
file as described in the Test::Perl::Critic documentation.  This is
implemented here as a more extensible variant of the explicit
configuration in the previous commit.