Use Class::Load to check for available test modules
... which allows one to avoid a string evaluation in order to check that
a given test module is available or not for the POD and Perl::Critic
tests.  Although this does add one extra dependency for the
distribution, it makes the test code adhere to the same Perl best
practices that the main code is held to.  This change also paves the way
for the test files to be checked as part of the `critic.t` tests.